pending: fix move queue

+++ Ball.cs
@@ -6,7 +6,7 @@ public class Ball : IObject
     public double SpeedX;
     public double SpeedY;
 
-    public Ball(int x, int y,double speedX, double speedY, IBehavior behavior, string? label = null) : base(x, y, Color.Green, behavior, label)
+    public Ball(double x, double y,double speedX, double speedY, IBehavior behavior, string? label = null) : base(x, y, Color.Green, behavior, label)
     {
         SpeedX = speedX;
         SpeedY = speedY;
